+++ mex/editor/merge/main.py
@@ -78,7 +78,10 @@ def entity_type_choice_merged(choice: tuple[str, bool]) -> rx.Component:
     return rx.checkbox(
         choice[0],
         checked=choice[1],
-        on_change=MergeState.set_entity_type_merged(choice[0]),
+        on_change=[
+            MergeState.set_entity_type_merged(choice[0]),
+            MergeState.refresh(["merged"]),
+        ],
         disabled=MergeState.is_loading,
     )
 
@@ -88,7 +91,10 @@ def entity_type_choice_extracted(choice: tuple[str, bool]) -> rx.Component:
     return rx.checkbox(
         choice[0],
         checked=choice[1],
-        on_change=MergeState.set_entity_type_extracted(choice[0]),
+        on_change=[
+            MergeState.set_entity_type_extracted(choice[0]),
+            MergeState.refresh(["extracted"]),
+        ],
         disabled=MergeState.is_loading,
     )